Change packages member for runtime vars to StrKIND, and initialize split strings
This commit changes the length of the packages member of runtime variables from
ShortStrKIND to StrKIND to accomodate longer package lists that exist for some
registry-defined variables.

This commit also modifies the mpas_split_string_new routine to initialize the
trailing characters for each string that is split out of the given delimited
string. This avoids an issue with garbage characters being present in the
substrings.